etl/requirements.txt
pandas
pyyaml
sqlalchemy
psycopg2
requests
cryptography